feat: abstract login form
