+++ pages/login.jsx
@@ -1,104 +1,27 @@
-import { gql, useMutation } from "@apollo/client";
 import {
   Box,
   Button,
   Divider,
   Flex,
-  FormControl,
-  FormErrorMessage,
-  FormLabel,
   Heading,
-  IconButton,
-  Input,
-  InputGroup,
-  InputRightElement,
   Link as ChakraLink,
   SimpleGrid,
-  Stack,
   Text,
   useColorModeValue,
-  useDisclosure,
-  useToast,
   VisuallyHidden,
 } from "@chakra-ui/react";
 import Head from "next/head";
 import Link from "next/link";
-import { useRouter } from "next/router";
-import { useForm } from "react-hook-form";
 import { FaFacebook, FaGoogle } from "react-icons/fa";
-import { HiEye, HiEyeOff } from "react-icons/hi";
-import LoadingSpinner from "../components/common/LoadingSpinner";
+import LoginForm from "../components/auth/LoginForm";
 import ThemeButton from "../components/common/ThemeButton";
-import { useAuth } from "../context/AuthContext";
-
-const LOG_IN_MUTATION = gql`
-  mutation logIn($email: String!, $password: String!) {
-    signInUser(input: { email: $email, password: $password }) {
-      accessToken
-      user {
-        id
-        email
-        name
-        profilePicture
-      }
-    }
-  }
-`;
 
 function Login() {
-  // next hooks
-  const router = useRouter();
-
-  // form hooks
-  const {
-    register,
-    handleSubmit,
-    formState: { errors },
-  } = useForm();
-
-  // auth hooks
-  const { dispatch } = useAuth();
-  const [signInUser, { loading }] = useMutation(LOG_IN_MUTATION);
-
-  // chakra hooks
-  const toast = useToast();
-  const { isOpen, onToggle } = useDisclosure();
-
   // chakra colors
   const textColors = useColorModeValue("purple.500", "purple.200");
   const bgColor = useColorModeValue("gray.50", "inherit");
   const boxColor = useColorModeValue("white", "gray.700");
   const separatorColor = useColorModeValue("gray.600", "gray.400");
-
-  const onSubmit = async ({ email, password }) => {
-    try {
-      const { data } = await signInUser({
-        variables: {
-          email,
-          password,
-        },
-      });
-
-      router.push("/");
-
-      dispatch({
-        type: "LOGIN",
-        payload: {
-          token: data.signInUser.accessToken,
-          user: data.signInUser.user,
-        },
-      });
-    } catch (error) {
-      toast({
-        title: error.message,
-        status: "error",
-        duration: 3000,
-        isClosable: true,
-      });
-    }
-  };
-
-  if (loading) return <LoadingSpinner />;
 
   return (
     <>
@@ -157,62 +80,7 @@ function Login() {
               sm: "lg",
             }}
           >
-            <form onSubmit={handleSubmit(onSubmit)}>
-              <Stack spacing={6}>
-                <FormControl id="email" isInvalid={!!errors.email}>
-                  <FormLabel>Email address</FormLabel>
-                  <Input
-                    type="email"
-                    {...register("email", {
-                      required: "Please enter a valid email address",
-                      pattern: !/^[A-Z0-9._%+-]+@[A-Z0-9.-]+\.[A-Z]{2,}$/i,
-                    })}
-                  />
-                  <FormErrorMessage>{errors?.email?.message}</FormErrorMessage>
-                </FormControl>
-
-                <FormControl id="password" isInvalid={!!errors.password}>
-                  <Flex justify="space-between">
-                    <FormLabel>Password</FormLabel>
-                    <ChakraLink
-                      color={textColors}
-                      fontWeight="semibold"
-                      fontSize="sm"
-                    >
-                      Forgot Password?
-                    </ChakraLink>
-                  </Flex>
-                  <InputGroup size="md">
-                    <Input
-                      pr="4.5rem"
-                      type={isOpen ? "text" : "password"}
-                      autoComplete="current-password"
-                      {...register("password", {
-                        required: "Please enter a password",
-                      })}
-                    />
-                    <InputRightElement>
-                      <IconButton
-                        bg="transparent !important"
-                        variant="ghost"
-                        aria-label={
-                          isOpen ? "Mask password" : "Reveal password"
-                        }
-                        icon={isOpen ? <HiEyeOff /> : <HiEye />}
-                        onClick={onToggle}
-                      />
-                    </InputRightElement>
-                  </InputGroup>
-                  <FormErrorMessage>
-                    {errors?.password?.message}
-                  </FormErrorMessage>
-                </FormControl>
-
-                <Button type="submit" colorScheme="purple">
-                  Log In
-                </Button>
-              </Stack>
-            </form>
+            <LoginForm />
 
             <Flex align="center" color="gray.300" mt="5">
               <Box flex="1">
